# docs/swagger.yaml
basePath: /api
...
host: localhost:5000
info:
  contact: {}
  description: 兽剧内容聚合平台后端服务 API 文档（neo-server，行为对齐 Express 版）
  title: 兽剧聚合平台 API
  version: 1.0.0
paths:
...
  /auth/cancel-deletion:
    post:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 取消注销申请
      tags:
      - 认证
...
  /auth/change-email:
    put:
      consumes:
      - application/json
      parameters:
      - description: newEmail/password
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      summary: 修改绑定邮箱（超管）
      tags:
      - 认证
  /auth/change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: currentPassword/newPassword
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 修改密码
      tags:
      - 认证
...
  /auth/confirm-device-login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 6 位登录码
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 用户对象或 need2FA
          schema:
            additionalProperties: true
            type: object
      summary: 确认设备登录
      tags:
      - 认证
  /auth/deletion-status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      summary: 注销申请状态
      tags:
      - 认证
  /auth/email-notification-prefs:
    put:
      consumes:
      - application/json
      parameters:
      - description: 7 个布尔偏好键
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      summary: 更新邮件通知偏好
      tags:
      - 认证
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: email/altcha
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 忘记密码
      tags:
      - 认证
...
  /auth/login-2fa:
    post:
      consumes:
      - application/json
      parameters:
      - description: email/twoFactorToken/twoFactorChallenge
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 用户对象
          schema:
            additionalProperties: true
            type: object
      summary: 2FA 登录
      tags:
      - 认证
...
  /auth/request-deletion:
    post:
      responses:
        "200":
          description: deletionRequestedAt/deleteAt
          schema:
            additionalProperties: true
            type: object
      security:
      - bearerAuth: []
      summary: 申请注销账号
      tags:
      - 认证
  /auth/request-email-change:
    post:
      consumes:
      - application/json
      parameters:
      - description: password/newEmail/altcha
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 申请修改邮箱
      tags:
      - 认证
  /auth/resend-verification:
    post:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - bearerAuth: []
      summary: 重发验证码（登录用户）
      tags:
      - 认证
  /auth/resend-verification-by-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: email/altcha
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 按邮箱重发验证码
      tags:
      - 认证
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: token/newPassword
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 重置密码
      tags:
      - 认证
...
  /auth/verify-device:
    post:
      consumes:
      - application/json
      parameters:
      - description: device-verify JWT
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: verified/loginCode
          schema:
            additionalProperties: true
            type: object
      summary: 设备验证：邮箱链接换取登录码
      tags:
      - 认证
  /auth/verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: 6 位验证码
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 验证邮箱
      tags:
      - 认证
  /auth/verify-email-change:
    post:
      consumes:
      - application/json
      parameters:
      - description: email-change JWT
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: email
          schema:
            additionalProperties: true
            type: object
      summary: 确认修改邮箱
      tags:
      - 认证
...
securityDefinitions:
  bearerAuth:
    description: 输入 JWT access token（不含 Bearer 前缀；亦可通过 accessToken cookie）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"